+++ src/app.module.ts
@@ -1,4 +1,9 @@
-import { Module } from '@nestjs/common';
+import {
+  MiddlewareConsumer,
+  Module,
+  NestModule,
+  RequestMethod,
+} from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AuthModule } from './auth/auth.module';
@@ -11,6 +16,7 @@ import { PostsModule } from './posts/posts.module';
 import { ThreadsModule } from './threads/threads.module';
 import { UsersModule } from './users/users.module';
 import { PrivateMessagesModule } from './private-messages/private-messages.module';
+import { RequestLoggingMiddleware } from './log/request.logging.middleware';
 
 @Module({
   imports: [
@@ -31,4 +37,10 @@ import { PrivateMessagesModule } from './private-messages/private-messages.modul
   controllers: [AppController],
   providers: [],
 })
-export class AppModule {}
+export class AppModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer
+      .apply(RequestLoggingMiddleware)
+      .forRoutes({ path: '*', method: RequestMethod.ALL });
+  }
+}