+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [2.2.5] - TODO
 
+### Changed
+
+- Request/response logging now utilizes both middlewares and interceptors to also log requests that were turned down (e.g. due to CORS).
+
 ### Chores
 
 - CORS origins are now being logged on startup.